Add empty_state_widget and refactor NullContentWidgets
