+++ packages/uni_app/lib/view/academic_path/schedule_page.dart
@@ -6,7 +6,7 @@ import 'package:uni/model/providers/riverpod/default_consumer.dart';
 import 'package:uni/model/providers/riverpod/lecture_provider.dart';
 import 'package:uni/view/academic_path/widgets/schedule_page_shimmer.dart';
 import 'package:uni/view/academic_path/widgets/schedule_page_view.dart';
-import 'package:uni_ui/common_widgets/EmptyStateWidget.dart';
+import 'package:uni_ui/common_widgets/empty_state_widget.dart';
 
 class SchedulePage extends ConsumerWidget {
   SchedulePage({super.key, DateTime? now}) : now = now ?? DateTime.now();
